pyproject.toml
[tool.poetry]
name = "pizza_itlab"
version = "0.1.0"
description = ""
authors = ["ilham <ilham@ilham.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
psycopg2 = "^2.9.3"
djangorestframework = "^3.13.1"
bcrypt = "^3.2.0"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
